+++ odoo/local-src/sf_stock/models/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 # Part of sensefly.
 import stock_location
+from . import stock_production_lot
+from . import stock_move